列索引 减少聚合扫描

+++ src/main/java/com/alibaba/lindorm/contest/lsm/IntIndexItem.java
@@ -12,7 +12,7 @@ public class IntIndexItem extends ColumnIndexItem {
         this.batchMax = batchMax;
     }
 
-    public double getBatchSum() {
+    public long getBatchSum() {
         return batchSum;
     }
